requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
scipy